+++ src/backends/elasticsearch/ElasticsearchException.ts
@@ -1,0 +1,9 @@
+export class ElasticsearchException implements Error {
+  constructor(name: string, s: string) {
+    this.name = name
+    this.message = s
+  }
+
+  message: string
+  name: string
+}
