+++ src/components/App.tsx
@@ -1,16 +1,16 @@
 import { Component, Fragment } from "inferno"
-import { Logs } from "./logs"
-import { Range } from "../helpers/time"
-import { AttachHistogramCallback, Histogram } from "./histogram"
-import { Logquacious } from "../services/logquacious"
-import { Query } from "../services/query"
-import { ChangeFilterCallback, FilterDropdown, FilterItem } from "./filterDropdown"
-import { ChangeSettingCallback, MenuSetting } from "./menuSetting"
-import { Direction, Theme } from "../services/prefs"
-import { Welcome } from "./welcome"
-import { Error } from "./error"
-import { Title } from "./title"
-import { SearchBar, SearchBarCallback } from "./searchBar"
+import { Logs } from "./Logs"
+import { Range } from "../helpers/Time"
+import { AttachHistogramCallback, Histogram } from "./Histogram"
+import { Logquacious } from "../services/Logquacious"
+import { Query } from "../services/Query"
+import { ChangeFilterCallback, FilterDropdown, FilterItem } from "./FilterDropdown"
+import { ChangeSettingCallback, MenuSetting } from "./MenuSetting"
+import { Direction, Theme } from "../services/Prefs"
+import { Welcome } from "./Welcome"
+import { Error } from "./Error"
+import { Title } from "./Title"
+import { SearchBar, SearchBarCallback } from "./SearchBar"
 import { Menu } from "./menu"
 import { ChangeRangeCallback, Picker } from "./picker/picker"
