+++ src/components/Histogram.tsx
